% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow.R
\name{arrow}
\alias{arrow}
\title{Insert Arrow emojis}
\usage{
arrow(direction)
}
\arguments{
\item{direction}{Character denoting the direction of the arrow. Should be one
of “up”, “up-right”, “right”, “down-right”, “down”, “down-left”, “left”,
“up-left”, “up-down”, or “left-right”.}
}
\description{
Insert Arrow emojis
}
\details{
This function is vectorized. Wrong input of \code{direction} will result in \code{NA}s.

#@return Character vector of emojis.
}
\examples{
arrow("up-down")

arrow(c("up", "up", "down", "down", "left", "right", "left", "right"))
}
