% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock.R
\name{clock}
\alias{clock}
\title{emoji version of time}
\usage{
clock(time)
}
\arguments{
\item{time}{a \code{POSIXct} object}
}
\value{
Character vector of emojis showing the closest time.
}
\description{
emoji version of time
}
\details{
This function is vectorized.
}
\examples{
times <- as.POSIXct("2021-09-17 14:33:21 PDT") + seq(1:30) * 3500
clock(times)
}
