% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moon.R
\name{day_in_synodic_cycle}
\alias{day_in_synodic_cycle}
\title{Calculate the number of days in the synodic month}
\usage{
day_in_synodic_cycle(date)
}
\arguments{
\item{date}{a date}
}
\description{
This uses the approximation described in the
\href{https://en.wikipedia.org/wiki/Lunar_phase}{lunar phase wikipedia page}, i.e.
the number of days since \code{1900/01/01} modulo the length of
a synodic month (\code{29.530588853} days)
}
\examples{
day_in_synodic_cycle(Sys.Date())
}
\keyword{internal}
