% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji-package.R
\docType{package}
\name{emoji-package}
\alias{emoji-package}
\title{emoji: Data and Function to Work with Emojis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Contains data about emojis with relevant metadata, and functions to work with emojis when they are in strings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://emilhvitfeldt.github.io/emoji/}
  \item \url{https://github.com/EmilHvitfeldt/emoji}
  \item Report bugs at \url{https://github.com/EmilHvitfeldt/emoji/issues}
}

}
\author{
\strong{Maintainer}: Emil Hvitfeldt \email{emilhhvitfeldt@gmail.com} (\href{https://orcid.org/0000-0002-0679-1945}{ORCID})

Other contributors:
\itemize{
  \item Hadley Wickham \email{hadley@rstudio.com} (Data parsing code from hadley/emo) [contributor]
  \item Romain François (Data parsing code from hadley/emo) [contributor]
}

}
\keyword{internal}
