% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rx.R
\name{emoji_detect}
\alias{emoji_detect}
\title{Detect the presence or absence of emojis in a string}
\usage{
emoji_detect(string)
}
\arguments{
\item{string}{Input vector. Either a character vector, or something coercible
to one}
}
\value{
A logical vector
}
\description{
Vectorised over \code{string}
}
\seealso{
\code{\link[stringr:str_detect]{stringr::str_detect()}}
}
