% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emoji.R
\name{emoji_find}
\alias{emoji_find}
\title{List all emoji with a given keyword}
\usage{
emoji_find(keyword)
}
\arguments{
\item{keyword}{Character, Emoji keyword.}
}
\description{
This function will look in \code{emoji_keyword} to report back the given emojis.
}
\examples{
emoji_find("happy")
emoji_find("cat")
emoji_find("family")
}
