% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emoji_keyword}
\alias{emoji_keyword}
\title{Emoji Keywords}
\format{
named list of characters with 7665 elements
}
\source{
Unicode® Full Emoji Charts v16.0 - https://www.unicode.org/emoji/charts-16.0/full-emoji-list.html

Unicode® Emoji Charts v16.0 - https://www.unicode.org/emoji/charts-16.0/emoji-list.html

Unicode® Emoji Ordering, v16.0 - https://www.unicode.org/emoji/charts/emoji-ordering.txt#' @source \url{https://github.com/github/gemoji}

\url{https://github.com/muan/emojilib}
}
\usage{
emoji_keyword
}
\description{
This list contains information about which emojis are contained in which
keywords.
}
\seealso{
emojis emoji_name
}
\keyword{datasets}
