% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rx.R
\name{emoji_locate}
\alias{emoji_locate}
\alias{emoji_locate_all}
\title{Locate the position of emojis in a string}
\usage{
emoji_locate(string)

emoji_locate_all(string)
}
\arguments{
\item{string}{Input vector}
}
\value{
For \code{emoji_locate} an integer matrix, for \code{emoji_locate_all} a list
of integer matrices
}
\description{
Vectorised over \code{string}
}
\examples{
string <- paste(c(letters[1:4], emoji_name[1:6]), collapse = " ")

emoji_locate(string)

emoji_locate_all(string)
}
