% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rx.R
\name{emoji_match}
\alias{emoji_match}
\alias{emoji_match_all}
\title{Extract emojis from a string}
\usage{
emoji_match(string)

emoji_match_all(string)
}
\arguments{
\item{string}{Input vector}
}
\value{
see \code{\link[stringr:str_match]{stringr::str_match()}}
}
\description{
Vectorized over \code{string}
}
\seealso{
\link[stringr:str_match]{stringr::str_match}
}
