% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emoji_modifiers}
\alias{emoji_modifiers}
\title{Emoji Modifiers}
\format{
tibble with 3 columns and
\code{nrow(emoji_modifiers)} rows
\describe{
\item{emoji_modifiers}{character representation of the emoji with modifiers}
\item{emoji}{character representation of the emoji without modifiers}
\item{modifiers}{list of modifiers}
}
}
\source{
\href{https://www.unicode.org/emoji/charts-15.0/full-emoji-list.html}{Unicode® Full Emoji Charts v15.0}

\href{https://www.unicode.org/emoji/charts-15.0/emoji-list.html}{Unicode® Emoji Charts v15.0}

\href{https://www.unicode.org/emoji/charts/emoji-ordering.txt}{Unicode® Emoji Ordering, v15.0}

\url{https://github.com/github/gemoji}

\url{https://github.com/muan/emojilib}
}
\usage{
emoji_modifiers
}
\description{
This data set contains all the emojis with modifiers, their unmodified
version as well as a list of the the modifiers.
}
\seealso{
emojis emoji_name
}
\keyword{datasets}
