% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rx.R
\docType{data}
\name{emoji_rx}
\alias{emoji_rx}
\title{A regular expression to catch all emojis}
\format{
character vector
}
\source{
\url{https://www.unicode.org/reports/tr51/#emoji_data}
}
\usage{
emoji_rx
}
\description{
This regex will capture all \code{fully-qualified} and \code{minimally-qualified}
emojis.
}
\keyword{datasets}
