% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{flag}
\alias{flag}
\title{Insert Flag Emojis}
\usage{
flag(name, return_key = FALSE)
}
\arguments{
\item{name}{Character denoting the place of the flag. Set \code{return_key = TRUE}
to get full list of allowed names.}

\item{return_key}{Logical, set to TRUE to get full list of allowed names.}
}
\value{
Character vector of emojis.
}
\description{
Insert Flag Emojis
}
\details{
This function is vectorized. The input is being normalized before matching
which will hopefully lead to lower friction and easier matching. Punctuation
is being removed and case is not taken into consideration when matching.
You can run \code{flag(return_key = TRUE)} to get full list of allowed names.
}
\examples{
flag(c("Vietnam", "Greenland", "Estonia", "Denmark", "united states"))

flag(c("US Virgin Islands", "U.S. Virgin Islands", "u.s. virgin islands"))
}
