% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keycap.R
\name{keycap}
\alias{keycap}
\title{Keycap emoji sequence}
\usage{
keycap(x)
}
\arguments{
\item{x}{character, must be a number between 0 and 10, "#", or "*".}
}
\value{
a keycap version of \code{x}
}
\description{
Keycap emoji sequence
}
\details{
This function is vectorized.
}
\examples{
keycap(6)
keycap('#')

keycap(1:10)
}
