% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medal.R
\name{medal}
\alias{medal}
\title{Insert medal emojis}
\usage{
medal(place)
}
\arguments{
\item{place}{Character denoting the place of the medal. See details for
allowed names.}
}
\description{
Insert medal emojis
}
\details{
This function is vectorized. There are a 1st, 2nd and 3rd place medals and
allowed names are listed below. Note that matches are made without case.

\itemize{
\item{1st place medal}{"1", "1st", or "gold"}
\item{2nd place medal}{"2", "2nd", or "silver"}
\item{3rd place medal}{"3", "3rd", or "bronze"}
}

#@return Character vector of emojis.
}
\examples{
medal(1:3)

medal("gold")
medal("Gold")
}
