% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moon.R
\name{moon}
\alias{moon}
\title{Insert Moon Phase Emoji}
\usage{
moon(date, day = day_in_synodic_cycle(date))
}
\arguments{
\item{date}{a date}

\item{day}{number of days since new moon}
}
\value{
a moon emoji
}
\description{
Insert Moon Phase Emoji
}
\details{
This function is vectorized.
If not supplied, \code{day} is calculated using the
approximation of \link{day_in_synodic_cycle}, i.e the number of
days since a known new moon modulo \code{29.530588853} days.
}
\examples{
moon(Sys.Date())

january <- as.Date("2021-01-01") + 0:30
moon(january)
}
