% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{shape}
\alias{shape}
\title{Insert Arrow emojis}
\usage{
shape(color, type)
}
\arguments{
\item{color}{Character, denoting the color of the shape. Must be one of "red",
"orange", "yellow", "green", "blue", "purple", "brown", "black", "white".}

\item{type}{Character, denoting the type of shape. Must be one of "heart",
"circle", or "square.}
}
\description{
Insert Arrow emojis
}
\details{
This function is vectorized.

#@return Character vector of emojis.
}
\examples{
shape("yellow", "heart")

shape("yellow", c("heart", "circle", "square"))

shape(color = c("red", "orange", "yellow", "green", "blue",
                "purple", "brown", "black", "white"),
      type = "circle")

outer(
  c("red", "orange", "yellow", "green", "blue",
    "purple", "brown", "black", "white"),
  c("heart", "circle", "square"),
  shape
)
}
