% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoo.R
\name{zoo}
\alias{zoo}
\title{Random Animals}
\usage{
zoo(size, replace = FALSE)
}
\arguments{
\item{size}{a non-negative integer giving the number of items to choose.}

\item{replace}{should sampling be with replacement? Defaults to \code{FALSE}.}
}
\value{
Character vector of animal emojis.
}
\description{
This function returns random animals emojis.
}
\examples{
set.seed(1234)

zoo(1)

zoo(10)
}
