% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rx.R
\name{emoji_count}
\alias{emoji_count}
\title{Count the number of emojis in a string}
\usage{
emoji_count(string)
}
\arguments{
\item{string}{Input vector}
}
\value{
An integer vector
}
\description{
Vectorised over \code{string}
}
\examples{
string <- paste(c(letters[1:4], emoji_name[1:6]), collapse = " ")

emoji_count(string)

emoji_count(emoji_name[1:6])
}
\seealso{
\code{\link[stringr:str_count]{stringr::str_count()}}
}
