% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emoji_keyword}
\alias{emoji_keyword}
\title{Emoji Keywords}
\format{
named list of characters with 6637 elements
}
\source{
\href{https://www.unicode.org/emoji/charts-13.1/full-emoji-list.html}{Unicode® Full Emoji Charts v13.1}

\href{https://www.unicode.org/emoji/charts-13.1/emoji-list.html}{Unicode® Emoji Charts v13.1}

\href{https://www.unicode.org/emoji/charts/emoji-ordering.txt}{Unicode® Emoji Ordering, v13.1}

\url{https://github.com/github/gemoji}

\url{https://github.com/muan/emojilib}
}
\usage{
emoji_keyword
}
\description{
This list contains information about which emojis are contained in which
keywords.
}
\seealso{
emojis emoji_name
}
\keyword{datasets}
