% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rx.R
\name{emoji_match}
\alias{emoji_match}
\alias{emoji_match_all}
\title{Extract matched emojis from a string}
\usage{
emoji_match(string)

emoji_match_all(string)
}
\arguments{
\item{string}{Input vector}
}
\value{
see \code{\link[stringr:str_match]{stringr::str_match()}}
}
\description{
Vectorized over \code{string}
}
\examples{
chars <- c(letters[1:4], emoji_name[1:6])

set.seed(1234)

strings <- lapply(1:10, function(x) paste(sample(chars, x), collapse = ""))

extracts <- emoji_match(strings)


extracts <- emoji_match_all(strings)
}
\seealso{
\link[stringr:str_match]{stringr::str_match}
}
