% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emoji_name}
\alias{emoji_name}
\title{Emoji Names}
\format{
named character vector with 4538 elements
}
\source{
\href{https://www.unicode.org/emoji/charts-13.1/full-emoji-list.html}{Unicode® Full Emoji Charts v13.1}

\href{https://www.unicode.org/emoji/charts-13.1/emoji-list.html}{Unicode® Emoji Charts v13.1}

\href{https://www.unicode.org/emoji/charts/emoji-ordering.txt}{Unicode® Emoji Ordering, v13.1}

\url{https://github.com/github/gemoji}

\url{https://github.com/muan/emojilib}
}
\usage{
emoji_name
}
\description{
This vector is a named vector of emojis, where then names are unique
descriptive identifiers for the emojis. This vector is well suited to be used
as a tool to replace emojis with natural language descriptions.
}
\details{
Some emojis will appear multiple times since they have multiple names
associated with them. Such as "grinning" and "grinning_face" leading to the
same emoji.
}
\seealso{
emojis emoji_keyboard
}
\keyword{datasets}
