% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rx.R
\name{emoji_replace}
\alias{emoji_replace}
\alias{emoji_replace_all}
\title{Replace emojis in a string}
\usage{
emoji_replace(string, replacement)

emoji_replace_all(string, replacement)
}
\arguments{
\item{string}{Input vector}

\item{replacement}{A character vector of replacements. Should either be
of length 1 or the same length as \code{string}. See \code{\link[stringr:str_replace]{stringr::str_replace()}}
for details}
}
\value{
A character vector
}
\description{
Vectorised over \code{string} and \code{replacement}
}
\examples{
emoji_replace(emoji_name[1], "_emoji_")

string <- paste(c(letters[1:4], emoji_name[1:6]), collapse = " ")

emoji_replace_all(emoji_name[1:6], "_emoji_")
}
