% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rx.R
\name{emoji_replace_name}
\alias{emoji_replace_name}
\title{Replace emojis in a string with name}
\usage{
emoji_replace_name(string)
}
\arguments{
\item{string}{Input vector}
}
\value{
A character vector
}
\description{
Vectorised over \code{string}
}
\details{
Each emoji is replaced with human readable string in the form
\verb{:name_of_emoji:}.
}
\examples{
example <- c(
  paste0("This is an emoji; ", emoji("person_facepalming")),
  paste0("You can write slides in ", emoji("key"), emoji("musical_note"))
)

example

emoji_replace_name(example)
}
